BDN-215. Bidding. Extracted ExecuteRoundUseCase
