+++ bidon/src/test/java/org/bidon/sdk/config/models/auctions/impl/AuctionImplTest.kt
@@ -61,8 +61,7 @@ internal class AuctionImplTest : ConcurrentTest() {
             adaptersSource = adaptersSource,
             getAuctionRequest = getAuctionRequestUseCase,
             statsRequest = statsRequestUseCase,
-            conductBiddingAuction = mockk(),
-            conductNetworkAuction = mockk()
+            executeRound = mockk()
         )
     }
